change truffle to hardhat
